Buildings dictionary replaced with function
The dictionary had the issue with the fact that it always output an identical class. For buildings the class saved to the map needs to be able to have different variables called that are specific to the building instances itself. This differs from professions which run in a similar manner. Professions don't need to be specific to the villager because the only time a villager never needs to call specific data from it. It's more of a tag attached to the villager rather than a building that has unique data.